+++ Watsonia.Data.TestPerformance/WatsoniaLinqTests.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Data.Entity;
+using Microsoft.EntityFrameworkCore;
 using Watsonia.Data.TestPerformance.Entities;
 
 namespace Watsonia.Data.TestPerformance
@@ -29,7 +29,7 @@ namespace Watsonia.Data.TestPerformance
 			return watch.ElapsedMilliseconds;
 		}
 
-		public long GetPlayerByID(int id)
+		public long GetPlayerByID(long id)
 		{
 			var watch = new Stopwatch();
 			watch.Start();
@@ -42,7 +42,7 @@ namespace Watsonia.Data.TestPerformance
 			return watch.ElapsedMilliseconds;
 		}
 
-		public long GetPlayersForTeam(int teamID)
+		public long GetPlayersForTeam(long teamID)
 		{
 			var watch = new Stopwatch();
 			watch.Start();
@@ -58,7 +58,7 @@ namespace Watsonia.Data.TestPerformance
 			return watch.ElapsedMilliseconds;
 		}
 
-		public long GetTeamsForSport(int sportID)
+		public long GetTeamsForSport(long sportID)
 		{
 			var watch = new Stopwatch();
 			watch.Start();
